docs(common-css): move Label and Busy Indicator to Patterns folder [ci visual]
