+++ packages/common-css/stories/patterns/busy-indicator/contrast-mode.example.html
