+++ packages/common-css/stories/patterns/busy-indicator/desktop.example.html
