+++ packages/common-css/stories/patterns/busy-indicator/extended-busy-indicator.example.html
