+++ packages/common-css/stories/patterns/busy-indicator/standard.example.html
