+++ packages/common-css/stories/patterns/label/default-example.example.html
